move compliance matrix prompt to the new folder
